Relax constraints on NonlinearProblem interface for computeRJ The const generics and 2D array slice constraints on the NonlinearProblem trait have caused some issues down the road in a number of crates and made certain designs quite difficult.
I've relaxed these constraints by moving over to 1D array slices as the input for all inputs which allows us to get rid of the const generic trait as well. Assuming we're dealing with contiguous data types we can easily make use of the bytemuck to cast between 1D and 2D slices without really suffering from any performance regressions in our benchmarks. I've added a few functions that automatically do the necessary bytemuck calls for users which makes things simple. They also have checks in place internally to make sure the slices are divisiable by the const generic type we want them to be.

Sadly even with these changes, we need to stay on nightly as const generics are still behind on stable for the features we need.